Makes this repo a package -- as makes sure the architectures are those in the paper
