+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "utils/pytorch_utils"]
-	path = utils/pytorch_utils
+[submodule "pointnet2/utils/pytorch_utils"]
+	path = pointnet2/utils/pytorch_utils
 	url = https://github.com/erikwijmans/etw_pytorch_utils.git
