+++ README.rst
@@ -14,4 +14,4 @@ Building CUDA kernels
 Exampling training
 ------------------
 
-Two training examples are provided by ``train_sem_seg.py`` and ``train_cls.py``.  The datasets for both will be downloaded automatically by default
+Two training examples are provided by ``pointnet2/train/train_sem_seg.py`` and ``train_cls.py``.  The datasets for both will be downloaded automatically by default
