+++ pointnet2/models/pointnet2_msg_cls.py
@@ -1,14 +1,9 @@
-import os, sys
-BASE_DIR = os.path.dirname(os.path.abspath(__file__))
-sys.path.append(BASE_DIR)
-sys.path.append(os.path.join(BASE_DIR, "../utils"))
-
 import torch
 import torch.nn as nn
-from torch.autograd import Variable
-import pytorch_utils as pt_utils
-from pointnet2_modules import PointnetSAModule, PointnetFPModule, PointnetSAModuleMSG
-from pointnet2_utils import RandomDropout
+from pointnet2.utils import pytorch_utils as pt_utils
+from pointnet2.utils.pointnet2_modules import (
+    PointnetSAModuleMSG, PointnetSAModule
+)
 from collections import namedtuple
 
 
@@ -18,8 +13,8 @@ def model_fn_decorator(criterion):
     def model_fn(model, data, epoch=0, eval=False):
         with torch.set_grad_enabled(not eval):
             inputs, labels = data
-            inputs = inputs.cuda(async=True)
-            labels = labels.cuda(async=True)
+            inputs = inputs.to('cuda', non_blocking=True)
+            labels = labels.to('cuda', non_blocking=True)
 
             preds = model(inputs)
             labels = labels.view(-1)
@@ -38,9 +33,9 @@ def model_fn_decorator(criterion):
     return model_fn
 
 
-class Pointnet2SSG(nn.Module):
+class Pointnet2MSG(nn.Module):
     r"""
-        PointNet2 with single-scale grouping
+        PointNet2 with multi-scale grouping
         Classification network
 
         Parameters
@@ -59,25 +54,33 @@ class Pointnet2SSG(nn.Module):
 
         self.SA_modules = nn.ModuleList()
         self.SA_modules.append(
-            PointnetSAModule(
+            PointnetSAModuleMSG(
                 npoint=512,
-                radius=0.2,
-                nsample=64,
-                mlp=[input_channels, 64, 64, 128],
+                radii=[0.1, 0.2, 0.4],
+                nsamples=[15, 32, 128],
+                mlps=[[input_channels, 32, 32,
+                       64], [input_channels, 64, 64, 128],
+                      [input_channels, 64, 96, 128]],
+                use_xyz=use_xyz
+            )
+        )
+
+        input_channels = 64 + 128 + 128
+        self.SA_modules.append(
+            PointnetSAModuleMSG(
+                npoint=128,
+                radii=[0.2, 0.4, 0.8],
+                nsamples=[32, 64, 128],
+                mlps=[[input_channels, 64, 64,
+                       128], [input_channels, 128, 128, 256],
+                      [input_channels, 128, 128, 256]],
                 use_xyz=use_xyz
             )
         )
         self.SA_modules.append(
             PointnetSAModule(
-                npoint=128,
-                radius=0.4,
-                nsample=64,
-                mlp=[128, 128, 128, 256],
-                use_xyz=use_xyz
+                mlp=[128 + 256 + 256, 256, 512, 1024], use_xyz=use_xyz
             )
-        )
-        self.SA_modules.append(
-            PointnetSAModule(mlp=[256, 256, 512, 1024], use_xyz=use_xyz)
         )
 
         self.FC_layer = nn.Sequential(
@@ -115,44 +118,3 @@ class Pointnet2SSG(nn.Module):
             xyz, features = module(xyz, features)
 
         return self.FC_layer(features.squeeze(-1))
-
-
-if __name__ == "__main__":
-    from torch.autograd import Variable
-    import numpy as np
-    import torch.optim as optim
-    import torch.autograd.profiler as profiler
-    B = 2
-    N = 2048
-    inputs = torch.randn(B, N, 6).cuda()
-    labels = torch.from_numpy(np.random.randint(0, 3, size=B)).cuda()
-    model = Pointnet2SSG(3, input_channels=3)
-    model.cuda()
-
-    optimizer = optim.Adam(model.parameters(), lr=1e-2)
-
-    print("testing with xyz")
-    model_fn = model_fn_decorator(nn.CrossEntropyLoss())
-    for _ in range(5):
-        optimizer.zero_grad()
-        _, loss, _ = model_fn(model, (inputs, labels))
-        loss.backward()
-        print(loss.data[0])
-        optimizer.step()
-
-    # use_xyz=False
-    inputs = torch.randn(B, N, 6).cuda()
-    labels = torch.from_numpy(np.random.randint(0, 3, size=B)).cuda()
-    model = Pointnet2SSG(3, input_channels=3, use_xyz=False)
-    model.cuda()
-
-    optimizer = optim.Adam(model.parameters(), lr=1e-2)
-
-    print("Testing without xyz")
-    model_fn = model_fn_decorator(nn.CrossEntropyLoss())
-    for _ in range(5):
-        optimizer.zero_grad()
-        _, loss, _ = model_fn(model, (inputs, labels))
-        loss.backward()
-        print(loss.data[0])
-        optimizer.step()
