+++ pointnet2/models/pointnet2_ssg_sem.py
@@ -1,0 +1,139 @@
+import torch
+import torch.nn as nn
+from pointnet2.utils import pytorch_utils as pt_utils
+from pointnet2.utils.pointnet2_modules import PointnetSAModule, PointnetFPModule
+from collections import namedtuple
+
+
+def model_fn_decorator(criterion):
+    ModelReturn = namedtuple("ModelReturn", ['preds', 'loss', 'acc'])
+
+    def model_fn(model, data, epoch=0, eval=False):
+        with torch.set_grad_enabled(not eval):
+            inputs, labels = data
+            inputs = inputs.to('cuda', non_blocking=True)
+            labels = labels.to('cuda', non_blocking=True)
+
+            preds = model(inputs)
+            loss = criterion(preds.view(labels.numel(), -1), labels.view(-1))
+
+            _, classes = torch.max(preds, -1)
+            acc = (classes == labels).float().sum() / labels.numel()
+
+            return ModelReturn(
+                preds, loss, {
+                    "acc": acc.item(),
+                    'loss': loss.item()
+                }
+            )
+
+    return model_fn
+
+
+class Pointnet2SSG(nn.Module):
+    r"""
+        PointNet2 with single-scale grouping
+        Semantic segmentation network that uses feature propogation layers
+
+        Parameters
+        ----------
+        num_classes: int
+            Number of semantics classes to predict over -- size of softmax classifier that run for each point
+        input_channels: int = 6
+            Number of input channels in the feature descriptor for each point.  If the point cloud is Nx9, this
+            value should be 6 as in an Nx9 point cloud, 3 of the channels are xyz, and 6 are feature descriptors
+        use_xyz: bool = True
+            Whether or not to use the xyz position of a point as a feature
+    """
+
+    def __init__(self, num_classes, input_channels=3, use_xyz=True):
+        super().__init__()
+
+        self.SA_modules = nn.ModuleList()
+        self.SA_modules.append(
+            PointnetSAModule(
+                npoint=1024,
+                radius=0.1,
+                nsample=32,
+                mlp=[input_channels, 32, 32, 64],
+                use_xyz=use_xyz
+            )
+        )
+        self.SA_modules.append(
+            PointnetSAModule(
+                npoint=256,
+                radius=0.2,
+                nsample=32,
+                mlp=[64, 64, 64, 128],
+                use_xyz=use_xyz
+            )
+        )
+        self.SA_modules.append(
+            PointnetSAModule(
+                npoint=64,
+                radius=0.4,
+                nsample=32,
+                mlp=[128, 128, 128, 256],
+                use_xyz=use_xyz
+            )
+        )
+        self.SA_modules.append(
+            PointnetSAModule(
+                npoint=16,
+                radius=0.8,
+                nsample=32,
+                mlp=[256, 256, 256, 512],
+                use_xyz=use_xyz
+            )
+        )
+
+        self.FP_modules = nn.ModuleList()
+        self.FP_modules.append(
+            PointnetFPModule(mlp=[128 + input_channels, 128, 128, 128])
+        )
+        self.FP_modules.append(PointnetFPModule(mlp=[256 + 64, 256, 128]))
+        self.FP_modules.append(PointnetFPModule(mlp=[256 + 128, 256, 256]))
+        self.FP_modules.append(PointnetFPModule(mlp=[512 + 256, 256, 256]))
+
+        self.FC_layer = nn.Sequential(
+            pt_utils.Conv1d(128, 128, bn=True), nn.Dropout(),
+            pt_utils.Conv1d(128, num_classes, activation=None)
+        )
+
+    def _break_up_pc(self, pc):
+        xyz = pc[..., 0:3].contiguous()
+        features = (
+            pc[..., 3:].transpose(1, 2).contiguous()
+            if pc.size(-1) > 3 else None
+        )
+
+        return xyz, features
+
+    def forward(self, pointcloud: torch.cuda.FloatTensor):
+        r"""
+            Forward pass of the network
+
+            Parameters
+            ----------
+            pointcloud: Variable(torch.cuda.FloatTensor)
+                (B, N, 3 + input_channels) tensor
+                Point cloud to run predicts on
+                Each point in the point-cloud MUST
+                be formated as (x, y, z, features...)
+        """
+        xyz, features = self._break_up_pc(pointcloud)
+
+        l_xyz, l_features = [xyz], [features]
+        for i in range(len(self.SA_modules)):
+            li_xyz, li_features = self.SA_modules[i](l_xyz[i], l_features[i])
+            l_xyz.append(li_xyz)
+            l_features.append(li_features)
+
+        for i in range(-1, -(len(self.FP_modules) + 1), -1):
+            l_features[i - 1] = self.FP_modules[i](
+                l_xyz[i - 1], l_xyz[i], l_features[i - 1], l_features[i]
+            )
+
+        return self.FC_layer(l_features[0]).transpose(1, 2).contiguous()
+
+
