+++ pointnet2/utils/__init__.py
@@ -1,0 +1,2 @@
+from . import pointnet2_utils
+from . import pointnet2_modules
