+++ pointnet2/utils/pointnet2_modules.py
@@ -2,8 +2,8 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
-import pointnet2_utils
-import pytorch_utils as pt_utils
+from pointnet2.utils import pointnet2_utils
+from pointnet2.utils import pytorch_utils as pt_utils
 from typing import List
